requirements.txt
pyrogram
tgcrypto
pyromod
pymongo
dnspython
flask
python-dotenv
pytz
motor
aiohttp
aiofiles
requests
beautifulsoup4
asyncio
ffmpeg-python
m3u8
cloudscraper
yt-dlp
pytube
pycryptodome
colorama
certifi
typing-extensions
